+++ src/main/java/de/canitzp/justabattery/JustAConfig.java
@@ -47,6 +47,8 @@ public class JustAConfig {
         public int battery_max_trace_width = 100;
         public String _allow_block_discharge = "The following value (true or false) defines if a battery is capable of extracting energy out of any energetic block.";
         public boolean allow_block_discharge = true;
+        public String _chargeup_creeper_energy_required = "How much energy is required to convert a creeper to a powered one. Use any negative number to disable this feature.";
+        public int chargeup_creeper_energy_required = 100_000;
     }
 
 }